backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-generativeai>=0.7.0
pydantic==2.9.2
python-dotenv>=1.0.0